Corrected usage of transactionId
Realex's documentation makes it look like the `<orderid>` tag for a refund should contain a new unique reference to the refund transaction. However, it turns out that actually, it requires it to be the id of our original payment. 
Therefore, I've corrected our terminology to bring it into line with other drivers that do refunds:
 - `transactionId` is now not used at all - Realex just doesn't care about our reference to this refund
 - A new required parameter called `originalTransactionId`. Other drivers don't seem to need this, so have essentially 'made up' the naming of this.